moved login functions to auth-setup.
Jira ticket: CAMS-401
